+++ src/applications/model/bulk-send-application.cc
@@ -17,6 +17,7 @@
 #include "ns3/simulator.h"
 #include "ns3/socket-factory.h"
 #include "ns3/socket.h"
+#include "ns3/tcp-socket-base.h"
 #include "ns3/tcp-socket-factory.h"
 #include "ns3/trace-source-accessor.h"
 #include "ns3/uinteger.h"
@@ -83,7 +84,12 @@ BulkSendApplication::GetTypeId()
             .AddTraceSource("TxWithSeqTsSize",
                             "A new packet is created with SeqTsSizeHeader",
                             MakeTraceSourceAccessor(&BulkSendApplication::m_txTraceWithSeqTsSize),
-                            "ns3::PacketSink::SeqTsSizeCallback");
+                            "ns3::PacketSink::SeqTsSizeCallback")
+            .AddTraceSource("TcpRetransmission",
+                            "The TCP socket retransmitted a packet",
+                            MakeTraceSourceAccessor(&BulkSendApplication::m_retransmissionTrace),
+                            "ns3::TcpSocketBase::RetransmissionCallback");
+
     return tid;
 }
 
@@ -185,6 +191,13 @@ BulkSendApplication::StartApplication() // Called at time specified by Start
         m_socket->SetConnectCallback(MakeCallback(&BulkSendApplication::ConnectionSucceeded, this),
                                      MakeCallback(&BulkSendApplication::ConnectionFailed, this));
         m_socket->SetSendCallback(MakeCallback(&BulkSendApplication::DataSend, this));
+        Ptr<TcpSocketBase> tcpSocket = DynamicCast<TcpSocketBase>(m_socket);
+        if (tcpSocket)
+        {
+            tcpSocket->TraceConnectWithoutContext(
+                "Retransmission",
+                MakeCallback(&BulkSendApplication::PacketRetransmitted, this));
+        }
     }
     if (m_connected)
     {
@@ -329,6 +342,17 @@ BulkSendApplication::DataSend(Ptr<Socket> socket, uint32_t)
         socket->GetPeerName(to);
         SendData(from, to);
     }
+}
+
+void
+BulkSendApplication::PacketRetransmitted(Ptr<const Packet> p,
+                                         const TcpHeader& header,
+                                         const Address& localAddr,
+                                         const Address& peerAddr,
+                                         Ptr<const TcpSocketBase> socket)
+{
+    NS_LOG_FUNCTION(this << p << header << localAddr << peerAddr << socket);
+    m_retransmissionTrace(p, header, localAddr, peerAddr, socket);
 }
 
 } // Namespace ns3