+++ src/applications/model/bulk-send-application.h
@@ -22,6 +22,8 @@ namespace ns3
 
 class Address;
 class Socket;
+class TcpHeader;
+class TcpSocketBase;
 
 /**
  * \ingroup applications
@@ -126,6 +128,14 @@ class BulkSendApplication : public Application
     /// Traced Callback: sent packets
     TracedCallback<Ptr<const Packet>> m_txTrace;
 
+    /// Traced Callback: retransmitted packets
+    TracedCallback<Ptr<const Packet>,
+                   const TcpHeader&,
+                   const Address&,
+                   const Address&,
+                   Ptr<const TcpSocketBase>>
+        m_retransmissionTrace;
+
     /// Callback for tracing the packet Tx events, includes source, destination,  the packet sent,
     /// and header
     TracedCallback<Ptr<const Packet>, const Address&, const Address&, const SeqTsSizeHeader&>
@@ -151,6 +161,20 @@ class BulkSendApplication : public Application
      * \param unused actually unused
      */
     void DataSend(Ptr<Socket> socket, uint32_t unused);
+
+    /**
+     *  \brief Packet retransmitted (called by TcpSocketBase sockets via callback)
+     *  \param p the retransmitted packet
+     *  \param header the TCP header
+     *  \param localAddr the local address
+     *  \param peerAddr the peer address
+     *  \param socket the socket that retransmitted the packet
+     */
+    void PacketRetransmitted(Ptr<const Packet> p,
+                             const TcpHeader& header,
+                             const Address& localAddr,
+                             const Address& peerAddr,
+                             Ptr<const TcpSocketBase> socket);
 };
 
 } // namespace ns3